// apps/backend/src/api/user/userModel.ts
import {extendZodWithOpenApi} from "@asteasolutions/zod-to-openapi";
import {z} from "zod";

import {commonValidations} from "@/common/utils/commonValidation";

extendZodWithOpenApi(z);

export type User = z.infer<typeof UserSchema>;
export const UserSchema = z.object({
    id: z.number(),
    uuid: z.string().uuid(),
    name: z.string(),
    email: z.string().email(),
    phone_number: z.string(),
    createdAt: z.date(),
    updatedAt: z.date(),
});

// Input Validation for 'GET users/:id' endpoint
export const GetUserSchema = z.object({
    params: z.object({id: commonValidations.id}),
});

export type UserWithPassword = User & { password: string };